client/requirements.txt
meshtastic
pypubsub
requests
flask
paho-mqtt
python-dotenv
